The store README's gate order carries the byte check, and the helper is said once
pi's MUST: `packages/store/README.md`'s enumerated gate order was factually
incomplete after this branch. It listed probe → optimistic check as steps 1–2
and gave `StaleWrite` one cause; the branch inserts a byte comparison between
them and makes the sentence three. `docs/architecture.md` got the full re-say
and the package README did not, which is the rule about the two shipping
together.

So the list grows a step 2 — the write's own paths compared by BYTES, why the
probe above cannot always do what it promises, what it costs, and that it is
the same distrust of stamps step 7 already has for this gate's own bytes on
the other side of the rename — and step 3 says its three causes.

The two "step N" cross-references become names ("the set the validate step
already judged", "the guarantee the decode step makes"). A numbered reference
into a list that just grew an item is a coupling nothing enforces, and this is
the edit that would have broken it.

pi's NITs, both taken:

- `replaceBehindTheStamps` had become two per-package copies of the same
  fixture, and what it carries is an INVARIANT rather than a convenience — the
  length guard is what caught #440's false-premise test, and a second copy is a
  second place for it to be softened. It moves to `@olai/store/testlib`, which
  is where the concept lives: what a stamp is made of, and therefore what can
  hide inside one, is this package's own. Same precedent and same reasoning as
  `@olai/format`'s `./testlib`, whose own package.json note names this exact
  failure ("the alternative was @olai/ops growing a byte-identical copy").
- the clean-path pin now says what it measures and what it only argues: the
  `decodes` assertion catches a check that widened to the whole tree, nothing
  here spies the disk so the per-write read count is argued from
  `probe.drifted`, and the guarantee that would really break is pinned
  functionally one test down.
